Add Meeting Aggregate root with Meeting initiation

+++ src/main/java/com/sama/calendar/application/Commands.kt
@@ -1,7 +1,7 @@
 package com.sama.calendar.application
 
 
-data class InitiateMeetingCommand(val duration: Int, val recipientEmail: String?)
+data class InitiateMeetingCommand(val duration: Long, val recipientEmail: String?)
 sealed class AddSuggestSlotCommand
 data class ModifySuggestSlotCommand(val slot: ProposedSlotDTO)
 data class RemoveSuggestSlotCommand(val slotId: Long)
